react issue #16582: how to avoid nested hooks example
