--- src/components/PlaceholderText/PlaceholderText.md
@@ -1,1 +1,0 @@
-## PlaceholderText
