--- src/components/PlaceholderText/index.js
@@ -1,5 +1,0 @@
-export {
-  default,
-  PlaceholderTextPropTypes,
-  PlaceholderTextDefaultProps
-} from "./PlaceholderText";
