+++ src/hooks/index.js
@@ -2,3 +2,4 @@ export { default as useDataAPI } from "./useDataAPI";
 export { default as useQuery } from "./useQuery";
 export { default as useData } from "./useData";
 export { default as useMarkdown } from "./useMarkdown";
+export { default as usePlaceholderTextRows } from "./usePlaceholderTextRows";
